refactor(script): split memorial buildins into buildin_md.c
Issue #61 eleventh step: move memorial dungeon helpers out of script.c.
