+++ src/map/script.c
@@ -6273,124 +6273,6 @@ int buildin_setfont(struct script_state *st)
 
 
 /*==========================================
- * メモリアルダンジョン作成
- *------------------------------------------
- */
-int buildin_mdcreate(struct script_state *st)
-{
-	struct map_session_data *sd = script_rid2sd(st);
-	char *memorial_name;
-	int party_id, ret;
-
-	nullpo_retr(0, sd);
-
-	memorial_name = conv_str(st,& (st->stack->stack_data[st->start+2]));
-
-	if(st->end>st->start+3)
-		party_id = conv_num(st,&(st->stack->stack_data[st->start+3]));
-	else
-		party_id = sd->status.party_id;
-
-	ret = memorial_create(memorial_name, party_id);
-	switch(ret) {
-		case MDCREATE_EXISTS:		// 既に生成済み
-			clif_msgstringtable2(sd, 0x52a, memorial_name);	// メモリアルダンジョン「%s」の予約が重複生成要請により失敗しました。
-			break;
-		case MDCREATE_PERMISSION:	// 権限がない
-			clif_msgstringtable2(sd, 0x529, memorial_name);	// メモリアルダンジョン「%s」の予約が権限問題により失敗しました。
-			break;
-		case MDCREATE_RESERVED:		// 既に予約済み
-			clif_msgstringtable2(sd, 0x528, memorial_name);	// メモリアルダンジョン「%s」の予約が予約重複により失敗しました。
-			break;
-		case MDCREATE_ERROR:		// その他エラー
-			clif_msgstringtable2(sd, 0x527, memorial_name);	// メモリアルダンジョン「%s」の予約に失敗しました。
-			break;
-	}
-
-	return 0;
-}
-
-/*==========================================
- * メモリアルダンジョン削除
- *------------------------------------------
- */
-int buildin_mddelete(struct script_state *st)
-{
-	struct map_session_data *sd = script_rid2sd(st);
-	int party_id = 0;
-
-	if(st->end>st->start+2)
-		party_id = conv_num(st,&(st->stack->stack_data[st->start+2]));
-	else if(sd)
-		party_id = sd->status.party_id;
-
-	if(party_id > 0) {
-		struct party *pt = party_search(party_id);
-		if(pt)
-			memorial_delete(pt->memorial_id);
-	}
-
-	return 0;
-}
-
-/*==========================================
- * メモリアルダンジョン入場
- *------------------------------------------
- */
-int buildin_mdenter(struct script_state *st)
-{
-	struct map_session_data *sd = script_rid2sd(st);
-	int ret = MDENTER_ERROR;
-	char *name = conv_str(st,& (st->stack->stack_data[st->start+2]));
-
-	if(sd) {
-		ret = memorial_enter(sd, name);
-	}
-
-	push_val(st->stack,C_INT,ret);
-
-	return 0;
-}
-
-/*==========================================
- * メモリアルダンジョン状態
- *------------------------------------------
- */
-int buildin_mdopenstate(struct script_state *st)
-{
-	struct map_session_data *sd = script_rid2sd(st);
-	int ret = -1;
-	char *name = conv_str(st,& (st->stack->stack_data[st->start+2]));
-
-	if(sd) {
-		ret = memorial_openstate(sd, name);
-	}
-
-	push_val(st->stack,C_INT,ret);
-
-	return 0;
-}
-
-/*==========================================
- * メモリアルダンジョンMAP名取得
- *------------------------------------------
- */
-int buildin_getmdmapname(struct script_state *st)
-{
-	char *str = conv_str(st,& (st->stack->stack_data[st->start+2]));
-	int id = script_getmemorialid(st);
-	int m = memorial_mapname2mapid(str, id);
-
-	if(m < 0)
-		push_str(st->stack,C_CONSTSTR,"");
-	else
-		push_str(st->stack,C_STR,(unsigned char *)aStrdup(map[m].name));
-
-	return 0;
-}
-
-
-/*==========================================
  * ClassからJob変換
  *------------------------------------------
  */
