+++ vcproj/map-server.vcxproj.filters
@@ -136,6 +136,9 @@
     <ClCompile Include="..\src\map\buildin_effect.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_md.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
       <Filter>map_sql</Filter>
     </ClCompile>
